Use PRIu64 from inttypes.h as the format specifier for uint64_t

+++ tests/unit/s2n_cbc_verify_test.c
@@ -15,6 +15,7 @@
 
 #include "s2n_test.h"
 
+#include <inttypes.h>
 #include <string.h>
 #include <stdio.h>
 #include <math.h>
@@ -203,11 +204,9 @@ int main(int argc, char **argv)
         int64_t hi = good_median + (3 * good_stddev);
 
         if ((int64_t) mac_median < lo || (int64_t) mac_median > hi) {
-            printf("\n\nRecord size: %d\nGood Median: %llu (Avg: %llu Stddev: %llu)\n"
-                   "Bad Median: %llu (Avg: %llu Stddev: %llu)\n\n", 
-                    i, (long long unsigned int)good_median, (long long unsigned int)good_avg,
-                    (long long unsigned int)good_stddev, (long long unsigned int)mac_median,
-                    (long long unsigned int)mac_avg, (long long unsigned int)mac_stddev);
+            printf("\n\nRecord size: %d\nGood Median: %" PRIu64 " (Avg: %" PRIu64 " Stddev: %" PRIu64 ")\n"
+                   "Bad Median: %" PRIu64 " (Avg: %" PRIu64 " Stddev: %" PRIu64 ")\n\n", 
+                    i, good_median, good_avg, good_stddev, mac_median, mac_avg, mac_stddev);
             FAIL();
         }
 
@@ -248,11 +247,9 @@ int main(int argc, char **argv)
         hi = good_median + (good_stddev);
 
         if ((int64_t) pad_median < lo || (int64_t) pad_median > hi) {
-            printf("\n\nRecord size: %d\nGood Median: %llu (Avg: %llu Stddev: %llu)\n"
-                   "Bad Median: %llu (Avg: %llu Stddev: %llu)\n\n", 
-                    i, (long long unsigned int)good_median, (long long unsigned int)good_avg,
-                    (long long unsigned int)good_stddev, (long long unsigned int)pad_median,
-                    (long long unsigned int)pad_avg, (long long unsigned int)pad_stddev);
+            printf("\n\nRecord size: %d\nGood Median: %" PRIu64 " (Avg: %" PRIu64 " Stddev: %" PRIu64 ")\n"
+                   "Bad Median: %" PRIu64 " (Avg: %" PRIu64 " Stddev: %" PRIu64 ")\n\n", 
+                    i, good_median, good_avg, good_stddev, pad_median, pad_avg, pad_stddev);
             FAIL();
         }
  
@@ -263,11 +260,9 @@ int main(int argc, char **argv)
         hi = mac_median + (mac_stddev / 2);
 
         if ((int64_t) pad_median < lo || (int64_t) pad_median > hi) {
-            printf("\n\nRecord size: %dMAC Median: %llu (Avg: %llu Stddev: %llu)\n"
-                   "PAD Median: %llu (Avg: %llu Stddev: %llu)\n\n", 
-                    i, (long long unsigned int)mac_median, (long long unsigned int)mac_avg,
-                    (long long unsigned int)mac_stddev, (long long unsigned int)pad_median,
-                    (long long unsigned int)pad_avg, (long long unsigned int)pad_stddev);
+            printf("\n\nRecord size: %dMAC Median: %" PRIu64 " (Avg: %" PRIu64 " Stddev: %" PRIu64 ")\n"
+                   "PAD Median: %" PRIu64 " (Avg: %" PRIu64 " Stddev: %" PRIu64 ")\n\n", 
+                    i, mac_median, mac_avg, mac_stddev, pad_median, pad_avg, pad_stddev);
             FAIL();
         }
     }
